chore(router): cleanup test code a bit, put strategy in outer loop
